Move TestRunner code into gdUnit4MonoApi
- simpify to run tests from a project by using test runner provided by the gdUnit4MonoApi assembly
