+++ PackageVersions.props
@@ -1,0 +1,8 @@
+<Project>
+  <PropertyGroup>
+    <GodotVersion>4.2.0-beta.5</GodotVersion>
+    <GodotNetSdkVersion>$(GodotVersion)</GodotNetSdkVersion>
+    <MicrosoftCodeAnalysisCSharpVersion>4.7.0</MicrosoftCodeAnalysisCSharpVersion>
+    <MoqVersion>4.18.4</MoqVersion>
+  </PropertyGroup>
+</Project>
