+++ api/gdUnit4MonoApi.csproj
@@ -1,4 +1,7 @@
-<Project Sdk="Godot.NET.Sdk/4.1.2">
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <Import Project="../PackageVersions.props" />
+
   <PropertyGroup>
     <Title>GdUnit4Mono</Title>
     <Version>4.2.0.0</Version>
@@ -37,9 +40,10 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.2.0" />
-    <PackageReference Include="Moq" Version="4.18.4" />
+    <PackageReference Include="GodotSharp" Version="$(GodotNetSdkVersion)" />
+    <PackageReference Include="Godot.SourceGenerators" Version="$(GodotNetSdkVersion)" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="$(MicrosoftCodeAnalysisCSharpVersion)" />
+    <PackageReference Include="Moq" Version="$(MoqVersion)" />
     <!-- Content Include=".gitignore" -->
     <Content Include="LICENSE" />
     <None Include="README.md" Pack="true" PackagePath="" />
