+++ gdUnit4Mono.csproj
@@ -1,12 +1,10 @@
-<Project Sdk="Godot.NET.Sdk/4.1.2">
+<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
     <LangVersion>11.0</LangVersion>
     <!--Force nullable warnings, you can disable if you want-->
     <Nullable>enable</Nullable>
-    <AssemblyName>GdUnit4Mono</AssemblyName>
     <RootNamespace>GdUnit4</RootNamespace>
-    <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
     <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
     <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
     <Deterministic>false</Deterministic>
