+++ gdUnit4Mono.sln
@@ -1,35 +1,30 @@
-﻿Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "gdUnit4Mono", "gdUnit4Mono.csproj", "{FCA65C1C-15F7-4936-A218-FDDDFD00ABCB}"
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "gdUnit4Mono", "gdUnit4Mono.csproj", "{690AC2F1-77C0-4D25-A3E7-C4ADF4DA9FF2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "gdUnit4MonoApi", "api\gdUnit4MonoApi.csproj", "{88F3D5AD-9730-4F48-857D-50DCBDCEBD09}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "gdUnit4MonoApi", "api\gdUnit4MonoApi.csproj", "{E5ED8330-DA73-49CE-A060-103CD6B8DC6D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "gdUnit4MonoTest", "test\gdUnit4MonoTest.csproj", "{4D0A33B6-C5A8-4861-99F4-A86400FF726F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "gdUnit4MonoTest", "test\gdUnit4MonoTest.csproj", "{B97C5043-B4BE-4156-BE0F-FDCBDECA61F6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
-		ExportDebug|Any CPU = ExportDebug|Any CPU
-		ExportRelease|Any CPU = ExportRelease|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FCA65C1C-15F7-4936-A218-FDDDFD00ABCB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FCA65C1C-15F7-4936-A218-FDDDFD00ABCB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FCA65C1C-15F7-4936-A218-FDDDFD00ABCB}.ExportDebug|Any CPU.ActiveCfg = ExportDebug|Any CPU
-		{FCA65C1C-15F7-4936-A218-FDDDFD00ABCB}.ExportDebug|Any CPU.Build.0 = ExportDebug|Any CPU
-		{FCA65C1C-15F7-4936-A218-FDDDFD00ABCB}.ExportRelease|Any CPU.ActiveCfg = ExportRelease|Any CPU
-		{FCA65C1C-15F7-4936-A218-FDDDFD00ABCB}.ExportRelease|Any CPU.Build.0 = ExportRelease|Any CPU
-		{88F3D5AD-9730-4F48-857D-50DCBDCEBD09}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{88F3D5AD-9730-4F48-857D-50DCBDCEBD09}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{88F3D5AD-9730-4F48-857D-50DCBDCEBD09}.ExportDebug|Any CPU.ActiveCfg = ExportDebug|Any CPU
-		{88F3D5AD-9730-4F48-857D-50DCBDCEBD09}.ExportDebug|Any CPU.Build.0 = ExportDebug|Any CPU
-		{88F3D5AD-9730-4F48-857D-50DCBDCEBD09}.ExportRelease|Any CPU.ActiveCfg = ExportRelease|Any CPU
-		{88F3D5AD-9730-4F48-857D-50DCBDCEBD09}.ExportRelease|Any CPU.Build.0 = ExportRelease|Any CPU
-		{4D0A33B6-C5A8-4861-99F4-A86400FF726F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4D0A33B6-C5A8-4861-99F4-A86400FF726F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4D0A33B6-C5A8-4861-99F4-A86400FF726F}.ExportDebug|Any CPU.ActiveCfg = ExportDebug|Any CPU
-		{4D0A33B6-C5A8-4861-99F4-A86400FF726F}.ExportDebug|Any CPU.Build.0 = ExportDebug|Any CPU
-		{4D0A33B6-C5A8-4861-99F4-A86400FF726F}.ExportRelease|Any CPU.ActiveCfg = ExportRelease|Any CPU
-		{4D0A33B6-C5A8-4861-99F4-A86400FF726F}.ExportRelease|Any CPU.Build.0 = ExportRelease|Any CPU
+		{E5ED8330-DA73-49CE-A060-103CD6B8DC6D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E5ED8330-DA73-49CE-A060-103CD6B8DC6D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E5ED8330-DA73-49CE-A060-103CD6B8DC6D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E5ED8330-DA73-49CE-A060-103CD6B8DC6D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B97C5043-B4BE-4156-BE0F-FDCBDECA61F6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B97C5043-B4BE-4156-BE0F-FDCBDECA61F6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B97C5043-B4BE-4156-BE0F-FDCBDECA61F6}.Release|Any CPU.ActiveCfg = Debug|Any CPU
+		{B97C5043-B4BE-4156-BE0F-FDCBDECA61F6}.Release|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 EndGlobal
