+++ global.json
@@ -1,0 +1,9 @@
+{
+    "sdk": {
+        "version": "7.0.0-rc",
+        "rollForward": "latestMinor"
+    },
+    "msbuild-sdks": {
+        "Godot.NET.Sdk": "4.2.0-beta.5"
+    }
+}
